Continue Command Revamping
I also set up two new macros that replaces the for loop for sending
something to all players to save time, since it's used in a ton of
places. Is it a good idea or not?

Also a commented a for loop in cmdListPlayers() to see if using the
C++11 equivalent of the win.udpPlayers loop works better after testing
the revamped command system.