Add Threads and Reorganize a bit
Console is done! The program compiles, but no GUI is present nor the
server actually starts, not even with the start command, but
mysteriously, the stop command works...

I believe this is because of the Widget class (which holds the core
server) isn't able to communicate with the Console class correctly
because of Threads. More signals and slots will need to be done I
presume.